refactor: extract environment handling
