+++ packages/viewer/src/routes/+layout.ts
@@ -4,36 +4,18 @@ import * as errors from '@/errors'
 import { sync } from '@/gateway'
 import type { LayoutLoad } from './$types'
 import { browser } from '$app/environment'
-
-import { HoneycombWebSDK } from '@honeycombio/opentelemetry-web'
-import { getWebAutoInstrumentations } from '@opentelemetry/auto-instrumentations-web'
+import { initializeHoneycomb } from '@/honeycomb'
+import { forHost } from '@/environment'
 
 export const load: LayoutLoad = async ({ fetch, url: { host } }) => {
-  const dev = host.match(/(dev(elopment)?[.-]|localhost)/) !== null
+  const env = forHost(host)
+  const dev = env === 'dev'
 
   if (browser) {
-    const configDefaults = {
-      ignoreNetworkEvents: true,
-    }
-
-    const sdk = new HoneycombWebSDK({
-      debug: dev,
-      apiKey: 'Sx608N12wwH9ZJ6qmHMVnA',
-      serviceName: 'π-base web',
-      instrumentations: [
-        getWebAutoInstrumentations({
-          '@opentelemetry/instrumentation-xml-http-request': configDefaults,
-          '@opentelemetry/instrumentation-fetch': configDefaults,
-          '@opentelemetry/instrumentation-document-load': configDefaults,
-        }),
-      ],
-    })
-    sdk.start()
+    initializeHoneycomb({ env })
   }
 
-  const errorHandler = dev
-    ? errors.log()
-    : errors.sentry({ environment: errorEnv(host) })
+  const errorHandler = dev ? errors.log() : errors.sentry({ env })
 
   const context = initialize({
     showDev: dev,
@@ -47,16 +29,4 @@ export const load: LayoutLoad = async ({ fetch, url: { host } }) => {
   await context.loaded()
 
   return context
-}
-
-function errorEnv(host: string): errors.Environment {
-  if (['topology.pi-base.org', 'topology.pages.dev'].includes(host)) {
-    return 'production'
-  }
-
-  if (host.includes('pages.dev')) {
-    return 'deploy-preview'
-  }
-
-  return 'dev'
 }
